requirements.txt
# AADNS - Advanced Autonomous Drone Navigation System
# Python 3.11+ required

# === CRITICAL: Colosseum AirSim client (NOT pip install airsim!) ===
# Install separately:
#   pip install git+https://github.com/CodexLabsLLC/Colosseum.git#subdirectory=PythonClient

# NumPy (MUST be < 2.0 for Colosseum compatibility)
numpy<2.0

# Deep Learning
torch>=2.0.0
torchvision>=0.17.0

# Computer Vision
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
scikit-image>=0.22.0

# Scientific computing
scipy>=1.12.0
filterpy>=1.4.5
pandas>=2.2.0

# Visualization
matplotlib>=3.8.0
pillow>=10.0.0

# Networking / Communication
requests>=2.31.0
pynng>=0.9.0

# Cryptography
cryptography>=42.0.0

# GUI
# tkinter - included with Python, no pip install needed
